moved src code into folder
